Write logic for rock_paper_scissor ui-version
I have written the logic for the second version that is gui version of
the rock-paper-scissor project of the foundations of TOP.

I have done a big mistake during this and because of that it took me 3
hours to write the logic which could have completed within 30 mins.
I directly jumped to my code editor before thinking about the logic to
solve this problem.. I started solving without getting the full picture
of the problem. and because of that I did not created a plan or steps to
solve this. I also forgot to commit any changes in between because I am
just doing anything which came to my mind like randomly..without
thinking about it.